// src/app/core/models/grounds.ts
export class Ground {
  id: any;
  groundNameAr!: string;
  groundNameEn!: string;
  purposeType!: number;
  ownerId: any;
  areaSize: any;
  readyForSale!: boolean;
  meterPrice: any;
  sellSpace: any;
  sellPrice: any;
  officeId: any;
  borderEast!: string;
  borderWest!: string;
  borderSouth!: string;
  borderNorth!: string;
  notes!: string;
  countryId: any;
  regionId: any;
  cityId: any;
  districtId: any;
  propertyInstrument!: string;
  addressInDetails!: string;
  previousOwnerNameAr!: string;
  previousOwnerNameEn!: string;
  noPlanned!: string;

}